Put lightweight tests into 1, the rest into 2
